change credit type to ulong, transactionsAmount to long, any id to ushort, and instalments counts to byte

+++ accounting/DataBase/DTOs/AccountDTO.cs
@@ -8,10 +8,10 @@ namespace accounting.DataBase.DTOs
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Key]
-        public int AccountId { get; set; }
+        public ushort AccountId { get; set; }
 
-        public int Credit { get; set; }
-        public int AvailableCredit { get; set; }
+        public ulong Credit { get; set; }
+        public ulong AvailableCredit { get; set; }
         public string CreateDate { get; set; }
         public string OwnerNationalId { get; set; }
         [ForeignKey("OwnerNationalId")] public PeoplesDTO Owner { get; set; }
