+++ accounting/DataBase/DTOs/LoanInstallmentsDTO.cs
@@ -7,11 +7,11 @@ namespace accounting.DataBase.DTOs
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Key]
-        public int Id { get; set; }
+        public ushort Id { get; set; }
 
-        public int LoanId { get; set; }
+        public ushort LoanId { get; set; }
         [ForeignKey("LoanId")] public LoansDTO Loan { get; set; }
-        public int Amount { get; set; }
+        public ulong Amount { get; set; }
         public string Date { get; set; }
     }
 }
