+++ accounting/Exceptions/NotEnoughAvailableCreditException.cs
@@ -4,8 +4,8 @@ namespace accounting.Exceptions
 {
     public class NotEnoughAvailableCreditException : Exception
     {
-        public int AvailableCredit;
-        public NotEnoughAvailableCreditException(int availableCredit) : base($"Not Enough credit, available credit: {availableCredit}")
+        public ulong AvailableCredit;
+        public NotEnoughAvailableCreditException(ulong availableCredit) : base($"Not Enough credit, available credit: {availableCredit}")
         {
             AvailableCredit = availableCredit;
         }
